Move component into Swift
