+++ compiler/core/src/decode.re
@@ -2,6 +2,43 @@ include Types;
 
 open Json.Decode;
 
+let parameterTypeMap =
+  [
+    ("text", Types.Reference("String")),
+    ("visible", Types.Reference("Boolean")),
+    ("numberOfLines", Types.Reference("Number")),
+    ("backgroundColor", Types.colorType),
+    ("image", Types.urlType),
+    /* Styles */
+    ("alignItems", Types.Reference("String")),
+    ("alignSelf", Types.Reference("String")),
+    ("flex", Types.Reference("Number")),
+    ("flexDirection", Types.Reference("String")),
+    ("font", Types.Reference("String")),
+    ("justifyContent", Types.Reference("String")),
+    ("marginTop", Types.Reference("Number")),
+    ("marginRight", Types.Reference("Number")),
+    ("marginBottom", Types.Reference("Number")),
+    ("marginLeft", Types.Reference("Number")),
+    ("paddingTop", Types.Reference("Number")),
+    ("paddingRight", Types.Reference("Number")),
+    ("paddingBottom", Types.Reference("Number")),
+    ("paddingLeft", Types.Reference("Number")),
+    ("borderRadius", Types.Reference("Number")),
+    ("width", Types.Reference("Number")),
+    ("height", Types.Reference("Number"))
+  ]
+  |> StringMap.fromList;
+
+
+let parameterType = (name) =>
+  switch (StringMap.find(name, parameterTypeMap)) {
+  | item => item
+  | exception Not_found =>
+    Js.log2("Unknown built-in parameter when deserializing:", name);
+    Reference("BuiltIn-Null")
+  };
+
 module Types = {
   let lonaType = (json) => {
     let referenceType = (json) => json |> string |> ((x) => Reference(x));
@@ -39,7 +76,7 @@ module Layer = {
       |> Js.Json.decodeObject
       |> Js.Option.getExn
       |> StringMap.fromJsDict
-      |> StringMap.mapi((key, value) => {ltype: Layer.parameterType(key), data: value});
+      |> StringMap.mapi((key, value) => {ltype: parameterType(key), data: value});
     {
       typeName: field("type", layerType, json),
       name: field("name", string, json),
