+++ compiler/core/src/layer.bs.js
@@ -9,7 +9,6 @@ var Pervasives                 = require("bs-platform/lib/js/pervasives.js");
 var Caml_string                = require("bs-platform/lib/js/caml_string.js");
 var Json_decode                = require("bs-json/src/Json_decode.js");
 var Logic$LonaCompilerCore     = require("./logic.bs.js");
-var Types$LonaCompilerCore     = require("./types.bs.js");
 var Caml_builtin_exceptions    = require("bs-platform/lib/js/caml_builtin_exceptions.js");
 var StringMap$LonaCompilerCore = require("./stringMap.bs.js");
 var StringSet$LonaCompilerCore = require("./stringSet.bs.js");
@@ -116,140 +115,6 @@ var LayerMap = /* module */[
   /* find_opt */find_opt
 ];
 
-var parameterTypeMap = StringMap$LonaCompilerCore.fromList(/* :: */[
-      /* tuple */[
-        "text",
-        /* Reference */Block.__(0, ["String"])
-      ],
-      /* :: */[
-        /* tuple */[
-          "visible",
-          /* Reference */Block.__(0, ["Boolean"])
-        ],
-        /* :: */[
-          /* tuple */[
-            "numberOfLines",
-            /* Reference */Block.__(0, ["Number"])
-          ],
-          /* :: */[
-            /* tuple */[
-              "backgroundColor",
-              Types$LonaCompilerCore.colorType
-            ],
-            /* :: */[
-              /* tuple */[
-                "image",
-                Types$LonaCompilerCore.urlType
-              ],
-              /* :: */[
-                /* tuple */[
-                  "alignItems",
-                  /* Reference */Block.__(0, ["String"])
-                ],
-                /* :: */[
-                  /* tuple */[
-                    "alignSelf",
-                    /* Reference */Block.__(0, ["String"])
-                  ],
-                  /* :: */[
-                    /* tuple */[
-                      "flex",
-                      /* Reference */Block.__(0, ["Number"])
-                    ],
-                    /* :: */[
-                      /* tuple */[
-                        "flexDirection",
-                        /* Reference */Block.__(0, ["String"])
-                      ],
-                      /* :: */[
-                        /* tuple */[
-                          "font",
-                          /* Reference */Block.__(0, ["String"])
-                        ],
-                        /* :: */[
-                          /* tuple */[
-                            "justifyContent",
-                            /* Reference */Block.__(0, ["String"])
-                          ],
-                          /* :: */[
-                            /* tuple */[
-                              "marginTop",
-                              /* Reference */Block.__(0, ["Number"])
-                            ],
-                            /* :: */[
-                              /* tuple */[
-                                "marginRight",
-                                /* Reference */Block.__(0, ["Number"])
-                              ],
-                              /* :: */[
-                                /* tuple */[
-                                  "marginBottom",
-                                  /* Reference */Block.__(0, ["Number"])
-                                ],
-                                /* :: */[
-                                  /* tuple */[
-                                    "marginLeft",
-                                    /* Reference */Block.__(0, ["Number"])
-                                  ],
-                                  /* :: */[
-                                    /* tuple */[
-                                      "paddingTop",
-                                      /* Reference */Block.__(0, ["Number"])
-                                    ],
-                                    /* :: */[
-                                      /* tuple */[
-                                        "paddingRight",
-                                        /* Reference */Block.__(0, ["Number"])
-                                      ],
-                                      /* :: */[
-                                        /* tuple */[
-                                          "paddingBottom",
-                                          /* Reference */Block.__(0, ["Number"])
-                                        ],
-                                        /* :: */[
-                                          /* tuple */[
-                                            "paddingLeft",
-                                            /* Reference */Block.__(0, ["Number"])
-                                          ],
-                                          /* :: */[
-                                            /* tuple */[
-                                              "borderRadius",
-                                              /* Reference */Block.__(0, ["Number"])
-                                            ],
-                                            /* :: */[
-                                              /* tuple */[
-                                                "width",
-                                                /* Reference */Block.__(0, ["Number"])
-                                              ],
-                                              /* :: */[
-                                                /* tuple */[
-                                                  "height",
-                                                  /* Reference */Block.__(0, ["Number"])
-                                                ],
-                                                /* [] */0
-                                              ]
-                                            ]
-                                          ]
-                                        ]
-                                      ]
-                                    ]
-                                  ]
-                                ]
-                              ]
-                            ]
-                          ]
-                        ]
-                      ]
-                    ]
-                  ]
-                ]
-              ]
-            ]
-          ]
-        ]
-      ]
-    ]);
-
 var stylesSet = Curry._1(StringSet$LonaCompilerCore.of_list, /* :: */[
       "alignItems",
       /* :: */[
@@ -299,20 +164,6 @@ var stylesSet = Curry._1(StringSet$LonaCompilerCore.of_list, /* :: */[
         ]
       ]
     ]);
-
-function parameterType(name) {
-  try {
-    return Curry._2(StringMap$LonaCompilerCore.find, name, parameterTypeMap);
-  }
-  catch (exn){
-    if (exn === Caml_builtin_exceptions.not_found) {
-      console.log("Unknown built-in parameter when deserializing:", name);
-      return /* Reference */Block.__(0, ["BuiltIn-Null"]);
-    } else {
-      throw exn;
-    }
-  }
-}
 
 function flatten(layer) {
   var inner = function (acc, layer) {
@@ -570,7 +421,7 @@ function getInsets(prefix, layer) {
           Caml_builtin_exceptions.match_failure,
           [
             "/Users/devin_abbott/Projects/ComponentStudio/ComponentStudio/compiler/core/src/layer.re",
-            185,
+            149,
             6
           ]
         ];
@@ -777,9 +628,7 @@ function toJavaScriptStyleSheetAST(layer) {
 }
 
 exports.LayerMap                      = LayerMap;
-exports.parameterTypeMap              = parameterTypeMap;
 exports.stylesSet                     = stylesSet;
-exports.parameterType                 = parameterType;
 exports.flatten                       = flatten;
 exports.find                          = find$1;
 exports.findByName                    = findByName;
