+++ compiler/core/src/logic.re
@@ -150,137 +150,3 @@ let rec toJavaScriptAST = (node) => {
   | None => Unknown
   }
 };
-
-let rec toSwiftAST = (colors, rootLayer: Types.layer, logicRootNode) => {
-  let identifierName = (node) =>
-    switch node {
-    | Identifier(ltype, [head, ...tail]) =>
-      switch head {
-      | "parameters" => Swift.Ast.SwiftIdentifier(List.hd(tail))
-      | "layers" =>
-        switch tail {
-        | [second, ...tail] when second == rootLayer.name =>
-          Swift.Ast.SwiftIdentifier(
-            List.tl(tail)
-            |> List.fold_left((a, b) => a ++ "." ++ Swift.Format.camelCase(b), List.hd(tail))
-          )
-        | [second, ...tail] =>
-          Swift.Ast.SwiftIdentifier(
-            tail
-            |> List.fold_left(
-                 (a, b) => a ++ "." ++ Swift.Format.camelCase(b),
-                 Swift.Format.layerName(second)
-               )
-          )
-        | _ => SwiftIdentifier("BadIdentifier")
-        }
-      | _ => SwiftIdentifier("BadIdentifier")
-      }
-    | _ => SwiftIdentifier("BadIdentifier")
-    };
-  let logicValueToSwiftAST = (x) =>
-    switch x {
-    | Identifier(_) => identifierName(x)
-    | Literal(value) => Swift.Document.lonaValue(colors, value)
-    | None => Empty
-    };
-  let typeAnnotationDoc =
-    fun
-    | Types.Reference(typeName) =>
-      switch typeName {
-      | "Boolean" => Swift.Ast.TypeName("Bool")
-      | _ => TypeName(typeName)
-      }
-    | Named(name, _) => TypeName(name);
-  let fromCmp = (x) =>
-    switch x {
-    | Types.Eq => "=="
-    | Neq => "!="
-    | Gt => ">"
-    | Gte => ">="
-    | Lt => "<"
-    | Lte => "<="
-    | Unknown => "???"
-    };
-  let unwrapBlock =
-    fun
-    | Block(body) => body
-    | node => [node];
-  let rec inner = (logicRootNode) =>
-    switch logicRootNode {
-    | Assign(a, b) =>
-      let (left, right) =
-        switch (logicValueToSwiftAST(b), logicValueToSwiftAST(a)) {
-        | (Swift.Ast.SwiftIdentifier(name), LiteralExpression(Boolean(value)))
-            when name |> Js.String.endsWith("visible") => (
-            Swift.Ast.SwiftIdentifier(name |> Js.String.replace("visible", "isHidden")),
-            Swift.Ast.LiteralExpression(Boolean(! value))
-          )
-        | (Swift.Ast.SwiftIdentifier(name), right) when name |> Js.String.endsWith("borderRadius") => (
-            Swift.Ast.SwiftIdentifier(
-              name |> Js.String.replace("borderRadius", "layer.cornerRadius")
-            ),
-            right
-          )
-        | (Swift.Ast.SwiftIdentifier(name), right) when name |> Js.String.endsWith("height") => (
-            Swift.Ast.SwiftIdentifier(
-              name |> Js.String.replace(".height", "HeightAnchorConstraint?.constant")
-            ),
-            right
-          )
-        | (Swift.Ast.SwiftIdentifier(name), right) when name |> Js.String.endsWith("width") => (
-            Swift.Ast.SwiftIdentifier(
-              name |> Js.String.replace(".width", "WidthAnchorConstraint?.constant")
-            ),
-            right
-          )
-        | nodes => nodes
-        };
-      Swift.Ast.BinaryExpression({"left": left, "operator": "=", "right": right})
-    | IfExists(a, body) =>
-      /* TODO: Once we support optional params, compare to nil or extract via pattern */
-      Swift.Ast.IfStatement({
-        "condition": logicValueToSwiftAST(a),
-        "block": unwrapBlock(body) |> List.map(inner)
-      })
-    | Block(body) => Swift.Ast.StatementListHelper(body |> List.map(inner))
-    | If(a, cmp, b, body) =>
-      Swift.Ast.IfStatement({
-        "condition":
-          Swift.Ast.BinaryExpression({
-            "left": logicValueToSwiftAST(a),
-            "operator": fromCmp(cmp),
-            "right": logicValueToSwiftAST(b)
-          }),
-        "block": unwrapBlock(body) |> List.map(inner)
-      })
-    | Add(lhs, rhs, value) =>
-      BinaryExpression({
-        "left": logicValueToSwiftAST(value),
-        "operator": "=",
-        "right":
-          Swift.Ast.BinaryExpression({
-            "left": logicValueToSwiftAST(lhs),
-            "operator": "+",
-            "right": logicValueToSwiftAST(rhs)
-          })
-      })
-    | Let(value) =>
-      switch value {
-      | Identifier(ltype, path) =>
-        Swift.Ast.VariableDeclaration({
-          "modifiers": [],
-          "pattern":
-            Swift.Ast.IdentifierPattern({
-              "identifier": List.fold_left((a, b) => a ++ "." ++ b, List.hd(path), List.tl(path)),
-              "annotation": Some(ltype |> typeAnnotationDoc)
-            }),
-          "init": (None: option(Swift.Ast.node)),
-          "block": (None: option(Swift.Ast.initializerBlock))
-        })
-      | _ => Empty
-      }
-    | None => Empty
-    };
-  logicRootNode |> unwrapBlock |> List.map(inner)
-};
